Added authorization through policies - added tests for authorization, deleting and updating

+++ app/Http/Controllers/ContactsController.php
@@ -5,35 +5,42 @@ namespace App\Http\Controllers;
 use App\Http\Resources\ContactResource;
 use Illuminate\Http\Request;
 use App\Models\Contact;
+use Illuminate\Validation\Rule;
 
 class ContactsController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Contact::class);
+
         return request()->user()->contacts;
     }
 
     public function store()
     {
-        Contact::create($this->validateData());
+        $this->authorize('create', Contact::class);
+
+        request()->user()->contacts()->create($this->validateData());
     }
 
     public function show(Contact $contact)
     {
-        if(request()->user()->id !== $contact->user_id) {
-            return abort(403);
-        }
+        $this->authorize('view', $contact);
 
         return new ContactResource($contact);
     }
 
     public function update(Contact $contact)
     {
+        $this->authorize('update', $contact);
+
         $contact->update($this->validateData());
     }
 
     public function destroy(Contact $contact)
     {
+        $this->authorize('delete', $contact);
+
         $contact->delete();
     }
